Combine matrix files
